Bug 544684 - Win64 issue, window handles are assumed to be 32-bit
2008-08-05  Tor Lillqvist  <tml@novell.com>

	Bug 544684 - Win64 issue, window handles are assumed to be 32-bit

	* gdk/win32/gdkcursor-win32.c
	* gdk/win32/xcursors.h: Change some gchar* to guchar* and vice
	versa to avoid gcc 4.4 signedness warnings.

	* gdk/win32/gdkevents-win32.c: Add some guchar and char pointer
	casts to get rid of gcc 4.4 signedness warnings. Print GdkAtom
	values in debugging output using the %p format.
	
	* gdk/win32/gdkkeys-win32.c
	* gdk/win32/gdkfont-win32.c
	* gdk/win32/gdkmain-win32.c: Add some casts to avoid gcc warnings.
 
	* gdk/win32/gdkwindow-win32.c: Use SetWindowLongPtr() instead of
	SetWindowLong().

	* gdk/win32/gdkwin32id.c (gdk_handle_hash): Use all 64 bits of a
	HANDLE on Win64.


svn path=/trunk/; revision=20994

+++ gdk/win32/gdkmain-win32.c
@@ -368,7 +368,7 @@ _gdk_win32_print_dc (HDC hdc)
 	   _gdk_win32_psstyle_to_string (extlogpen.elpPenStyle),
 	   _gdk_win32_psendcap_to_string (extlogpen.elpPenStyle),
 	   _gdk_win32_psjoin_to_string (extlogpen.elpPenStyle),
-	   extlogpen.elpWidth,
+	   (int) extlogpen.elpWidth,
 	   _gdk_win32_lbstyle_to_string (extlogpen.elpBrushStyle));
   g_print ("rop2: %s textcolor=%06lx\n",
 	   _gdk_win32_rop2_to_string (GetROP2 (hdc)),
